feat(whir): Add comments about code refactoring
