+++ package.json
@@ -13,6 +13,7 @@
     "@tailwindcss/vite": "^4.2.2",
     "react": "^19.2.4",
     "react-dom": "^19.2.4",
+    "react-router": "^7.13.2",
     "react-router-dom": "^7.13.2",
     "tailwindcss": "^4.2.2"
   },
