+++ src/App.jsx
@@ -1,6 +1,11 @@
-import { BrowserRouter as Router, Routes, Route, Navigate } from 'react-router-dom';
-import Login from "./pages/Login";
-import Dashboard from "./pages/Dashboard";
+import {
+  BrowserRouter as Router,
+  Routes,
+  Route,
+  Navigate,
+} from "react-router-dom";
+import Login from "./features/Login";
+import Dashboard from "./features/Dashboard";
 
 function App() {
   return (
